pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transparent-overlay"
version = "2.6.1"

authors = [
  { name = "Ilya Yakovenko", email = "ilya.a.yakovenko@gmail.com" },
]
description = "High-performance transparent graphics overlay for Windows"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
keywords = ["overlay", "graphics", "transparent", "windows", "gdi"]
dependencies = [
    "numpy>=1.19.0",
    "Pillow>=8.0.0",
    "pywin32>=300; sys_platform == 'win32'",
]

[project.optional-dependencies]
speedup = ["numba>=0.56"]
examples = ["psutil>=5.8.0", "pyautogui>=0.9.0"]


[project.license]
text = "MIT"